refactor: shared assets 디렉토리 구조 업데이트
