+++ packages/shared/src/assets/index.ts
@@ -1,2 +1,3 @@
-export { default as CheckIcon } from "./CheckIcon";
-export { default as GoogleIcon } from "./GoogleIcon";
+export { default as CheckIcon } from "./icon/Table/CheckIcon";
+export { default as GoogleIcon } from "./icon/Login/GoogleIcon";
+export { default as KakaoIcon } from "./icon/Login/KakaoIcon";
